Moved status endpoint
