+++ backend/src/data_entry/repository.rs
@@ -104,6 +104,28 @@ WHERE polling_station_id = ?
 
         Ok(())
     }
+    
+    pub async fn statuses(
+        &self,
+        election_id: u32,
+    ) -> Result<Vec<PollingStationDataEntry>, sqlx::Error> {
+        query_as!(
+            PollingStationDataEntry,
+            r#"
+SELECT
+  id AS "id: u32",
+  de.state AS "state: _",
+  updated_at
+FROM polling_stations AS p
+LEFT JOIN polling_station_data_entries AS de ON de.polling_station_id = p.id
+WHERE election_id = $1
+"#,
+            election_id
+        )
+        .fetch_all(&self.0)
+        .await
+    }
+
 }
 
 pub struct PollingStationResultsEntries(SqlitePool);
